One file changed, two added for about page work
-about.css added. This file contains the css code for the tabs and content on the about page
could not get the blue from template.css to transfer over, kept showing as grey. So current
blue is slightly different shade, will need to be fixed.

-tabs.js added. Contains javascript code for moving between tabs. Kept in seperate file because
it got a bit long and to keep about.html clean, just importing it when about.html is opened.

-updated about.html. Added three tab sections (may end up more or possibly one less). Only
one with content right now is Vision tab with our vision statement. The others will need content
added. Commented out the log in division as I believe we said were moving it to the menu bar.
Did not delete because wasn't 100% sure.