Updates
I have updated the header to move the sign up and the login to the right
side of the page. Also, the sign up and log in have a bar separating
them now. In the template I moved the old code to the bottom of the page
so that it is easier to get to the code that is needed.