updated ability to search for user
when logged in, can now search for user and will take to a new page that
will display search results. currently using a like with wildcards so
will display any user that contains what is searched, but if for example
testUser is spelled tostUser, will not return proper result. Still work
in progress