Fixed image upload and iteration 3
Granted permissions to users, groups, and others, to read and write to
the userImages folder. Might need to change permissions later. This
fixed the image uploads, so uploading event and profile photos works
now. Messed around with popup.css to try and change font. Fixed the
title and date of IterationReport3.