Linked checkboxes to photos
Linked the checkboxes for removing photos to the photos using labels.
Changed image alts to show what number the image is; checkable by
inspecting element. Added a function to display an alert when a photo is
uploaded successfully; needs to be implemented upon SUCCESSFUL upload -
might be hard to do so just left it in for now. Next step is to remove
checkboxes and images if the remove button is clicked.