Working on individualizing editPhotos
Hopefully the back end should work, but cannot currently test it. Right
now the buttons aren't displayed on editPhotos because something in the
php for photo removal isn't working right anymore. Need to work on 450
and 470 for the rest of the weekend, so I'll get to fixing it Monday.
Only change to createEvent was formatting.