Directories are created
Directories for user photos are created upon entering photo edit. Might
want to do this when they log in instead. Still having problems with
photo removal. The buttons do not display because the code never reaches
past the forms. Could be due to the php within a form that calls php.
Might have to move the inner php to the ppRemove and epRemove files.
Will test more later.