Added initial edits for setting profile photos
Added ppSet.php for setting the user's profile picture. Added a popup
form in editPhotos.php for setting the photo; we can add it to
profile.php too and it should work the same. Added styling for the new
popup, and added javascript to show/hide it. Need to add database
functionality to ppSet.php to put the image into the database; may also
need to create a separate folder called "current" or something in
userImages in order to hold the user's profile photo (unless I just do
it by the directory stored under current in userImages/profile/).