Back end for set current photo working ish
Back end for setting a user's current photo will store the photo in the
correct folder and in the database. However, for some reason AND
statements aren't working on the UPDATE statement in ppSet.php so all of
the rows for the images table will have the "current" image url.