Working on removing current photos
Current photo does not update to default if you remove the profile photo
that you are using as your current photo. I'm trying to remove the photo
from current if it has the same name as the photo being removed from
profile photos, then make the blank profile photo the current photo.